Use "libkdumpfile" when building "drgn"
We've integrated changes to "drgn" such that it will automatically
detect "libkdumpfile" if it's installed, and enable support for kernel
core dump files when it's detected. To ensure "drgn" is built with this
functionality enabled, we need to install the "libkdumpfile" package
after building it, and prior to building the "drgn" package.